sma-matching-backend/requirements.txt
# ===============================
# Framework Flask
# ===============================
Flask==2.3.3
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0

# ===============================
# MongoDB
# ===============================
pymongo==4.4.1

# ===============================
# JWT pour authentification
# ===============================
PyJWT==2.8.0
passlib[bcrypt]==1.7.4  # Pour hash des mots de passe

# ===============================
# Upload / parsing fichiers CV
# ===============================
python-docx==0.8.11   # Pour lire fichiers .docx
PyPDF2==3.1.0         # Pour lire fichiers PDF

# ===============================
# Crew AI SDK / requêtes
# ===============================
requests==2.31.0       # Pour appels HTTP Crew AI
